Time: 358 ms (5.02%), Space: 59.1 MB (64.21%) - LeetHub

+++ 2214-minimum-health-to-beat-game/2214-minimum-health-to-beat-game.cpp
@@ -1,5 +1,5 @@
 class Solution {
-    bool check(vector<int>& arr, int armor, long long health){
+    bool check(vector<int>& arr, int armor, long health){
         health -= max(arr[0]-armor, 0);
         for(int i=1; i<arr.size(); i++){
             health -= arr[i];
@@ -10,10 +10,9 @@ public:
     long long minimumHealth(vector<int>& damage, int armor) {
         sort(damage.begin(), damage.end(), greater<int>());
         
-        long long low = 0, high = accumulate(damage.begin(), damage.end(), 0ll)+1, ans = -1;
+        long low = 0, high = accumulate(damage.begin(), damage.end(), 0ll)+1, ans = -1;
         while(low <= high){
-            long long mid = (low+high)/2;
-            // cout << mid << " ";
+            long mid = (low+high)/2;
             if(check(damage, armor, mid))
                 ans = mid, high = mid-1;
             else
